+++ src/main/java/com/github/fishio/MenuField.java
@@ -2,6 +2,7 @@ package com.github.fishio;
 
 import java.util.ArrayList;
 
+import com.github.fishio.factories.EnemyFishFactory;
 import com.github.fishio.game.GameThread;
 
 import javafx.scene.canvas.Canvas;
