+++ src/main/java/com/github/fishio/PlayingField.java
@@ -11,6 +11,7 @@ import java.util.concurrent.ConcurrentLinkedQueue;
 import javafx.scene.canvas.Canvas;
 
 import com.github.fishio.behaviours.IMoveBehaviour;
+import com.github.fishio.factories.EnemyFishFactory;
 import com.github.fishio.game.GameThread;
 import com.github.fishio.gui.Renderer;
 import com.github.fishio.logging.Log;
