+++ src/test/java/com/github/fishio/TestPowerUpSpawner.java
@@ -14,6 +14,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mockito.Mockito;
 
+import com.github.fishio.factories.PowerUpSpawner;
 import com.github.fishio.game.GameThread;
 import com.github.fishio.gui.SlimGuiTest;
 import com.github.fishio.power_ups.PowerUp;
